Normalize error management and clear `dist/`
